ESEMM-8: Ensure Contribution exchange rate is always set
